requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
pydantic-settings==2.2.1
python-dotenv==1.0.1
langchain==0.1.11
faiss-cpu==1.8.0
sentence-transformers==2.5.1
requests==2.31.0
beautifulsoup4==4.12.3
pytest==8.0.2
httpx==0.27.0
python-multipart==0.0.9
loguru==0.7.2
